+++ CHANGELOG.rst
@@ -3,12 +3,13 @@ Change Log
 TODO
 ----
 
-0.3.3
+0.4.0
 ------------------
 * vistock/util.py: Added the support of .TWO to decide_market_color_style
 * vistock/util.py: Added doctest to decide_market_color_style
 * vistock/mpl/rsi.py: Created ta module and replaced talib package with ta
   module
+* Add features of IBD RS and IBD RS Rating
 
 0.3.2 [2024-07-22]
 ------------------